Add Fitur Edit Kriteria

+++ delete.php
@@ -13,14 +13,21 @@ if($item == 'kriteria'){
     $del = mysqli_query($conn,"DELETE FROM daftar_kriteria where id = '$id'");
 }
 if($item == 'lokasi'){
-    $del = mysqli_query($conn,"DELETE FROM tempat_wisata_tb where id = '$id'");
+    $result = mysqli_query($conn,"SELECT * from tempat_wisata_tb");
+    $rowcount=mysqli_num_rows($result);
+    if($rowcount == 1){
+        $message = "[GAGAL!!!] - Setidaknya harus ada 1 data lokasi wisata yang tersimpan pada database.";
+        echo "<script>alert('$message'); window.location.replace('data_lokasi_wisata.php');</script>";
+    }else{
+        $del = mysqli_query($conn,"DELETE FROM tempat_wisata_tb where id = '$id'");
 
-    $arr_kriteria = mysqli_query($conn,"SELECT * from daftar_kriteria_static");
-    while($data = mysqli_fetch_array($arr_kriteria)):
-        $tname = "fuzzy_";
-        $tname.= strtolower($data['kriteria']);
-        $del = mysqli_query($conn,"DELETE FROM {$tname} where id = '$id'");
-    endwhile;
+        $arr_kriteria = mysqli_query($conn,"SELECT * from daftar_kriteria_static");
+        while($data = mysqli_fetch_array($arr_kriteria)):
+            $tname = "fuzzy_";
+            $tname.= strtolower($data['kriteria']);
+            $del = mysqli_query($conn,"DELETE FROM {$tname} where id = '$id'");
+        endwhile;
+    }
 }
 
 if($del)
@@ -30,7 +37,9 @@ if($del)
         header("location:admin_page.php"); // redirects to all records page
     }
     if($item == 'lokasi'){
-        header("location:data_lokasi_wisata.php"); // redirects to all records page
+        $message = "Berhasil menghapus data lokasi wisata.";
+        echo "<script>alert('$message'); window.location.replace('data_lokasi_wisata.php');</script>";
+        //header("location:data_lokasi_wisata.php"); // redirects to all records page
     }
     exit;	
 }
